requirements.txt
google-adk>=1.0.0
litellm>=1.40.0
python-dotenv>=1.2.0
openai>=1.30.0
qdrant-client>=1.17.0
sentence-transformers>=5.3.0
transformers>=4.51.0
numpy>=1.26.0
torch>=2.0.0
rank-bm25>=0.2.2
huggingface-hub>=0.20.0
